new_branch commit 3

+++ app/src/main/java/com/hfad/allweather/domain/model/forecast_weather/CurrentForecastWeather.kt
@@ -2,7 +2,7 @@ package com.hfad.allweather.domain.model.forecast_weather
 
 data class CurrentForecastWeather(
     val condition: ConditionForecastWeather,
-    val temp_c: Int,
+    val temp_c: Double,
     val humidity: Int,
     val wind_kph: Double
 )
